backend/requirements.txt
django>=3.1,<3.2
djangorestframework
django-cors-headers